Added files and created structure for page design
I added the basic structure for the form page design, the second design that was given. This is just the basic structure and will need to be filled out to completely match the design. I also added the files to use the address-schema.js into your form so you don't need to use the one from the 686-674 as a dependency.